Squashed commit of the following:
commit 7049336e5ccc47812089078387c4dde190df18f4
Author: Paril <paril@alteredsoftworks.com>
Date:   Tue Sep 20 05:10:58 2016 -0400

    Collapse changes into one commit

# Conflicts:
#	Sources/Objectively/Class.h
#	Sources/Objectively/JSONPath.c
#	Sources/Objectively/Locale.c
#	Sources/Objectively/MutableString.c
#	Sources/Objectively/Types.h

+++ Sources/Objectively/MutableDictionary.c
@@ -178,13 +178,13 @@ static void setObjectForKey_resize(Dictionary *dict) {
 
 	if (dict->capacity) {
 
-		const float load = dict->count / dict->capacity;
+		const float load = dict->count / (float)dict->capacity;
 		if (load >= MUTABLEDICTIONARY_MAX_LOAD) {
 
 			size_t capacity = dict->capacity;
 			ident *elements = dict->elements;
 
-			dict->capacity = dict->capacity * MUTABLEDICTIONARY_GROW_FACTOR;
+			dict->capacity = (size_t)(dict->capacity * MUTABLEDICTIONARY_GROW_FACTOR);
 			dict->count = 0;
 
 			dict->elements = calloc(dict->capacity, sizeof(ident));
